Rename chiasma_deposit --> chiasmadeposit
The module name must match the repo name as given by user.
